% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{anonymise}
\alias{anonymise}
\title{Randomly anonymise a vector

This function randomly anonymises a vector, changing the values to letter codes.}
\usage{
anonymise(x, anon.level = 4, return.lookup = F)
}
\arguments{
\item{x}{The vector to be anonymised.}

\item{anon.level}{How many letters should be used in the anonymised codes.}

\item{return.lookup}{Should the function return a lookup table to find the original values?}
}
\value{
A vector the same length as the original, with values replaced by a randomly generated code. Identical values correspond to identical codes. If return.lookup = T, a table is also returned that allows the original values to be extracted from the anonymised list.
}
\description{
Randomly anonymise a vector

This function randomly anonymises a vector, changing the values to letter codes.
}
\note{
If the number of unique values in x is very large (exceeds 26^anon.level) the app will return an error. The function may also take a long time to run if the number of unique values in x is very high.
}
\examples{
## Anonymise the ShipNames in the data
anonymise(catcheffort.sample$ShipName)

## Keep a lookup table for later checking
ship.anon <- anonymise(catcheffort.sample$ShipName, return.lookup = T)
ship.anon$anon.list # List of anonymised values
ship.anon$anon.lookup # Lookup table
}
