% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{check.column.names}
\alias{check.column.names}
\title{Return missing column names}
\usage{
check.column.names(input.frame, column.target)
}
\arguments{
\item{input.frame}{Data frame to check.}

\item{column.target}{List of column names that should be in the target data frame.}
}
\value{
List of columns that are missing in the input frame.
}
\description{
The check.column.names function helps data checking by returning a list of column names that are missing
}
