% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{simulator.ui}
\alias{simulator.ui}
\title{Build the simulator ui}
\usage{
simulator.ui(masterdata)
}
\value{
A shiny ui object.
}
\description{
The simulator.ui function returns the ui object for the app, assuming the user has returned length data. It could be used to customise the tool for individual purposes.
}
\details{
As the function is built into the surveySimulator() call, some aspects of the function rely on things specified in that function. If you want to edit this for your own purposes, ensure you have checked all of the inputs!
}
