% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{surveySimulator}
\alias{surveySimulator}
\title{Run the Survey Simulation Tool}
\usage{
surveySimulator(catcheffort = NA, length = NA)
}
\arguments{
\item{catcheffort}{Data frame containing combined catch and effort data}

\item{length}{Optional data frame containing length data for the same geographical area as the catch and effort data}
}
\value{
Running an RShiny app, no data is returned to the R session by default.
}
\description{
The function to run the tool, which will open in a new window through RShiny.
}
\details{
Catcheffort has to be a data frame containing columns for Species, Catch, Effort (eg hooks), Gear, Nationality, ShipName, Latitude, Longitude and Season as a minimum. Running data through the simulatorData() function allows checking and name changing of columns to ensure it reaches minimum standards.
Length data is optional, and allows viewing of simulated length results in the app.
}
\examples{
## Run data using the included (randomly generated) data.
# Not specifying data runs the app in test mode
# Note that this data is randomly generated from mathematical distributions and is centered around longitude 0, latitude 0. It has no basis in any real CCAMLR dataset.

surveySimulator()
}
\seealso{
simulatorData()
}
